feat: add user, site, and site group detail retrieval functions
